Refactor file system
