+++ data-analysis-service/app/analysis/visuals/visualizations.py
@@ -1,0 +1,72 @@
+from app.analysis.visuals.churn_risk import create_churn_risk_visualization
+from app.analysis.visuals.client_value import create_client_value_visualization
+from app.analysis.visuals.credit_score import create_credit_score_visualization
+from app.analysis.visuals.loans import create_loan_recommendation_visualization
+
+
+def create_client_insights_visualization(insights_data):
+    """Create a comprehensive dashboard for all client insights"""
+    if not insights_data:
+        return None
+
+    # Create a grid of visualizations
+    credit_score_viz = create_credit_score_visualization(insights_data['credit_score'])
+    client_value_viz = create_client_value_visualization(insights_data['client_value'])
+    churn_risk_viz = create_churn_risk_visualization(insights_data['churn_risk'])
+    loan_rec_viz = create_loan_recommendation_visualization(insights_data['loan_recommendations'])
+
+    viz_html = ""
+
+    if credit_score_viz:
+        viz_html += f"""
+            <div class="card">
+                <h2>Credit Score</h2>
+                {credit_score_viz['gauge']}
+            </div>
+            <div class="card">
+                <h2>Score Components</h2>
+                {credit_score_viz['components']}
+            </div>
+        """
+
+    if client_value_viz:
+        viz_html += f"""
+            <div class="card">
+                <h2>Value Metrics</h2>
+                {client_value_viz['radar']}
+            </div>
+            <div class="card">
+                <h2>Total Value Score</h2>
+                {client_value_viz['gauge']}
+            </div>
+        """
+
+    if churn_risk_viz:
+        viz_html += f"""
+            <div class="card">
+                <h2>Churn Risk Score</h2>
+                {churn_risk_viz['gauge']}
+            </div>
+            <div class="card">
+                <h2>Risk Components</h2>
+                {churn_risk_viz['indicators']}
+            </div>
+        """
+
+    if loan_rec_viz:
+        viz_html += f"""
+            <div class="card">
+                <h2>Loan Recommendations</h2>
+                {loan_rec_viz['recommendations']}
+            </div>
+            <div class="card">
+                <h2>Likelihood to Repay</h2>
+                {loan_rec_viz['gauge']}
+            </div>
+            <div class="card">
+                <h2>Loan Application Flow</h2>
+                {loan_rec_viz['funnel']}
+            </div>
+        """
+
+    return viz_html
